requirements.txt
Django==1.8.12
djangorestframework==3.3.3
et-xmlfile==1.0.1
jdcal==1.2
mysql-connector-python==2.1.3
openpyxl==2.3.5
Pillow==3.2.0
